third_party/blink/web_tests/platform/generic/fast/forms/fieldset/fieldset-name-expected.txt

This test checks fieldset.name attribute.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Read from HTML
PASS fs1.name is "a"
PASS fs1.getAttribute("name") is "a"
fs1.name="b"
PASS fs1.name is "b"
PASS fs1.getAttribute("name") is "b"
fs1.name=1234
PASS fs1.name is "1234"
PASS fs1.getAttribute("name") is "1234"
fs1.name=null
PASS fs1.name is "null"
PASS fs1.getAttribute("name") is "null"
PASS fs1.hasAttribute("name") is true
fs1.name=undefined
PASS fs1.name is "undefined"
PASS fs1.getAttribute("name") is "undefined"
PASS fs1.hasAttribute("name") is true
PASS successfullyParsed is true

TEST COMPLETE
